[frontend] Install puma for production deployment
The current setup uses passenger from the host as the application
server, but puma is the standard application server for rails. This
could be handled externally, but adding it to the Gemfile makes the
whole process much more controlled.

Sadly, puma 3 is needed as newer versions switch to using nio4r. That
requires ruby 2.5, which is incompatible with rails 4.